Split DNS record validation into a separate module
